Move view data provider registration into webview
