+++ src/webviews/extension-side/serviceRegistry.node.ts
@@ -5,14 +5,20 @@
 import { IExtensionSingleActivationService, IExtensionSyncActivationService } from '../../platform/activation/types';
 import { IWebviewViewProvider, IWebviewPanelProvider } from '../../platform/common/application/types';
 import { IServiceManager } from '../../platform/ioc/types';
-import { IVariableViewProvider } from './variablesView/types';
+import { INotebookWatcher, IVariableViewProvider } from './variablesView/types';
 import { VariableViewActivationService } from './variablesView/variableViewActivationService';
 import { VariableViewProvider } from './variablesView/variableViewProvider';
 import { WebviewPanelProvider } from './webviewPanels/webviewPanelProvider';
 import { WebviewViewProvider } from './webviewViews/webviewViewProvider';
 import { JupyterVariableDataProvider } from './dataviewer/jupyterVariableDataProvider';
 import { JupyterVariableDataProviderFactory } from './dataviewer/jupyterVariableDataProviderFactory';
-import { IJupyterVariableDataProvider, IJupyterVariableDataProviderFactory } from './dataviewer/types';
+import {
+    IDataViewer,
+    IDataViewerDependencyService,
+    IDataViewerFactory,
+    IJupyterVariableDataProvider,
+    IJupyterVariableDataProviderFactory
+} from './dataviewer/types';
 import { INotebookExporter, INotebookImporter } from '../../kernels/jupyter/types';
 import { JupyterExporter } from './import-export/jupyterExporter.node';
 import { JupyterImporter } from './import-export/jupyterImporter.node';
@@ -22,17 +28,42 @@ import { RendererCommunication } from './plotView/rendererCommunication.node';
 import { PlotSaveHandler } from './plotView/plotSaveHandler.node';
 import { PlotViewHandler } from './plotView/plotViewHandler.node';
 import { DataViewerCommandRegistry } from './dataviewer/dataViewerCommandRegistry';
+import { DataViewer } from './dataviewer/dataViewer.node';
+import { IPlotViewer, IPlotViewerProvider } from './plotting/types';
+import { PlotViewer } from './plotting/plotViewer.node';
+import { DataViewerDependencyService } from './dataviewer/dataViewerDependencyService.node';
+import { PlotViewerProvider } from './plotting/plotViewerProvider.node';
+import { DataViewerFactory } from './dataviewer/dataViewerFactory';
+import { NotebookWatcher } from './variablesView/notebookWatcher';
 
 export function registerTypes(serviceManager: IServiceManager, _isDevMode: boolean) {
     serviceManager.addSingleton<IExtensionSingleActivationService>(IExtensionSingleActivationService, ServerPreload);
-    serviceManager.addSingleton<PlotSaveHandler>(PlotSaveHandler, PlotSaveHandler);
-    serviceManager.addSingleton<PlotViewHandler>(PlotViewHandler, PlotViewHandler);
     serviceManager.addSingleton<IExtensionSingleActivationService>(
         IExtensionSyncActivationService,
         RendererCommunication
     );
+
+    serviceManager.add<IDataViewer>(IDataViewer, DataViewer);
+    serviceManager.addSingleton<IDataViewerFactory>(IDataViewerFactory, DataViewerFactory);
+    serviceManager.add<IPlotViewer>(IPlotViewer, PlotViewer);
+    serviceManager.addSingleton<IPlotViewerProvider>(IPlotViewerProvider, PlotViewerProvider);
+    serviceManager.addSingleton<PlotSaveHandler>(PlotSaveHandler, PlotSaveHandler);
+    serviceManager.addSingleton<PlotViewHandler>(PlotViewHandler, PlotViewHandler);
+
+    serviceManager.addSingleton<IDataViewerDependencyService>(
+        IDataViewerDependencyService,
+        DataViewerDependencyService
+    );
+    serviceManager.addSingleton<IExtensionSingleActivationService>(
+        IExtensionSingleActivationService,
+        DataViewerCommandRegistry
+    );
+
     serviceManager.add<IWebviewViewProvider>(IWebviewViewProvider, WebviewViewProvider);
     serviceManager.add<IWebviewPanelProvider>(IWebviewPanelProvider, WebviewPanelProvider);
+
+    // Variable View
+    serviceManager.addSingleton<INotebookWatcher>(INotebookWatcher, NotebookWatcher);
     serviceManager.addSingleton<IExtensionSingleActivationService>(
         IExtensionSingleActivationService,
         VariableViewActivationService
@@ -43,12 +74,10 @@ export function registerTypes(serviceManager: IServiceManager, _isDevMode: boole
         IJupyterVariableDataProviderFactory,
         JupyterVariableDataProviderFactory
     );
+
+    // Import/Export
     serviceManager.add<INotebookExporter>(INotebookExporter, JupyterExporter);
     serviceManager.add<INotebookImporter>(INotebookImporter, JupyterImporter);
-    serviceManager.addSingleton<IExtensionSingleActivationService>(
-        IExtensionSingleActivationService,
-        DataViewerCommandRegistry
-    );
     serviceManager.addSingleton<IExtensionSingleActivationService>(
         IExtensionSingleActivationService,
         ExportCommandRegistry
